+++ FF/data/ffjson.c
@@ -216,7 +216,7 @@ static int hdlEsc(ffparser *p, int *st, int ch)
 	char buf[8];
 	int r;
 
-	if (p->esc[0] >= FFSLEN("uXXXX"))
+	if ((uint)p->esc[0] >= FFSLEN("uXXXX"))
 		return FFPARS_EESC; //too large escape sequence
 
 	p->esc[(uint)++p->esc[0]] = (char)ch;
